// packages/src/model/types/resource-model.ts
export interface ResCrudModelOptions {
  rowKey?: string
  initialPage?: number
  initialPageSize?: number
  pageSizeOptions?: number[]
  selectableRows?: boolean
  detailValueFormatter?: (value: unknown, key: string) => unknown
}

export interface ResReportModelOptions {
  initialPage?: number
  initialPageSize?: number
  pageSizeOptions?: number[]
}

export interface ResCrudModel {
  kind: 'crud'
  title: string
  rowKey: string
  pageSizeOptions: number[]
  selectableRows: boolean
  queryFields: ResourceCrudQueryField[]
  columns: ResourceCrudColumn[]
  formSections: ResourceCrudFormSection[]
  controller: ResourceCrudController
  detailFields: ComputedRef<ResourceCrudDetailField[]>
}

export interface ResReportModel {
  kind: 'report'
  title: string
  pageSizeOptions: number[]
  queryFields: ResourceCrudQueryField[]
  columns: ReportViewResultColumn[]
  controller: ReportViewController
}

export type ResModel = ResCrudModel | ResReportModel